chore: Restructured the main.rs into different files
